cortstim UI updates
