+++ package.json
@@ -6,7 +6,9 @@
   "scripts": {
     "start": "node server/server.js",
     "dev": "nodemon --experimental-json-modules server/server.js",
-    "docs:src": "typedoc --out docs/srcdoc src --mode file --theme ./node_modules/typedoc-neo-theme/bin/default --name WebFM Documentation",
+    "docs:src_file": "typedoc --out docs/srcdoc src --mode file --name WebFM Documentation",
+    "docs:src_modules": "typedoc --out docs/srcdoc src --mode modules --name WebFM Documentation",
+    "docs:json": "typedoc --json docs/srcdoc.json --mode modules",
     "docs:server": "cd docs && make html"
   },
   "nodemonConfig": {
@@ -38,12 +40,12 @@
   "license": "SEE LICENSE IN LICENSE.md",
   "sideEffects": false,
   "dependencies": {
-    "@babel/core": "^7.10.4",
+    "@babel/core": "^7.10.5",
     "@babel/plugin-proposal-class-properties": "^7.10.4",
     "@babel/plugin-proposal-export-default-from": "^7.10.4",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
     "@babel/plugin-transform-modules-commonjs": "^7.10.4",
-    "@babel/plugin-transform-runtime": "^7.10.4",
+    "@babel/plugin-transform-runtime": "^7.10.5",
     "@babel/preset-env": "^7.10.4",
     "@babel/preset-react": "^7.10.4",
     "@babel/preset-typescript": "^7.10.4",
@@ -56,11 +58,12 @@
     "clean-webpack-plugin": "^3.0.0",
     "compression": "^1.7.4",
     "css-loader": "^3.6.0",
+    "d3": "^5.16.0",
     "d3-axis": "^1.0.12",
     "d3-horizon-chart": "0.0.6",
     "d3-scale": "^3.2.1",
     "d3-scale-chromatic": "^1.5.0",
-    "d3-selection": "^1.4.1",
+    "d3-selection": "^1.4.2",
     "d3-shape": "^1.3.7",
     "express": "^4.17.1",
     "formidable": "^1.2.2",
@@ -71,7 +74,7 @@
     "postcss-loader": "^3.0.0",
     "precss": "^4.0.0",
     "react": "^16.13.1",
-    "react-bootstrap": "^1.1.1",
+    "react-bootstrap": "^1.3.0",
     "react-dom": "^16.13.1",
     "react-input-slider": "^5.1.7",
     "react-router-dom": "^5.2.0",
@@ -82,14 +85,15 @@
     "three": "^0.118.3",
     "ts-loader": "^7.0.5",
     "typedoc-neo-theme": "^1.0.9",
-    "typescript": "^3.9.6",
-    "webpack": "^4.43.0",
+    "typescript": "^3.9.7",
+    "webpack": "^4.44.0",
     "webpack-dev-middleware": "^3.7.2",
-    "webpack-merge": "^5.0.8",
+    "webpack-merge": "^5.0.9",
     "write-file-webpack-plugin": "^4.5.1"
   },
   "devDependencies": {
     "jsdoc-babel": "^0.5.0",
-    "typedoc-webpack-plugin": "^1.1.4"
+    "typedoc-webpack-plugin": "^1.1.4",
+    "nodemon": "2.0.4"
   }
 }
