+++ src/Components/Brain.tsx
@@ -1,101 +1,101 @@
-//@ts-nocheck
-import React, { useContext, useEffect, useState } from "react";
-import { Image } from "react-bootstrap";
-import { Context } from "../Context";
-import { fetchAndStoreBrain, fetchAndStoreGeometry } from "../shared/loaders";
-
-
+// Brain.tsx
 /**
- * ```
- * // Or you can specify the language explicitly
- * const instance = new Brain();
- * ```
- * @param props  Comment for parameter ´text´.
- * @returns      Comment for special return value.
+ * This is the doc comment for file1.ts
+ * @packageDocumentation
  */
-
+import React, { useContext, useEffect, useState, ReactSVGElement, useLayoutEffect } from "react";
+import { Context } from "../Context";
+import { Image } from 'react-bootstrap'
 
 const Brain = (props) => {
-  let {
-    subject,
-    setNewBrainCoord,
-    setNewBrain,
-    brain,
-    brainCoord,
-  }: any = useContext(Context);
-  if (subject.name == "") subject.name = "PY17N005";
-  const [brainCoords, setBrainCoords] = useState({});
-  const [dots, setDots] = useState<Element>();
-  let actualCoords = {};
+  let { subject, setNewSubject }: any = useContext(Context);
+  let [electrodes, setElectrodes] = useState()
+  let [img, setImg] = useState<string>()
+  const [size, setSize] = useState([]);
 
-  //Initial render once all the subjects are queried
+
+
+  useLayoutEffect(() => {
+    function updateSize() {
+      setSize([document.getElementById('container').offsetWidth, document.getElementById('container').offsetHeight]);
+    }
+    window.addEventListener('resize', updateSize);
+    updateSize();
+    return () => window.removeEventListener('resize', updateSize);
+  }, []);
   useEffect(() => {
-    let brainContainer = document.getElementById("brain");
+    (async () => {
+      console.log(subject)
+      let response11 = await fetch(`/api/geometry/${subject.name}`);
+      let geometry = await response11.json();
+      let response = await fetch(`/api/brain/${subject.name}`);
+      let brainRes = await response.arrayBuffer();
+      setNewSubject({
+        name: subject.name,
+        geometry
+      })
+      let binary = "";
+      let bytes = [].slice.call(new Uint8Array(brainRes));
+      bytes.forEach((b: any) => (binary += String.fromCharCode(b)));
+      setImg(binary)
+      let actualCoords = {};
 
-    Promise.all([
-      fetchAndStoreBrain(subject.name),
-      fetchAndStoreGeometry(subject.name),
-    ]).then((vals) => {
-      setNewBrain(vals[0]);
-      let geometry = vals[1];
 
-      Object.keys(geometry).forEach((electrodes) => {
-        let vals = geometry[electrodes];
+      Object.keys(geometry).forEach(electrodes => {
         actualCoords[electrodes] = {
-          u: vals.u * brainContainer.offsetWidth,
-          v: (1 - vals.v) * brainContainer.offsetHeight,
-          location: vals.location,
+          u: geometry[electrodes].u,
+          v: (1 - geometry[electrodes].v),
+          location: geometry[electrodes].location,
         };
       });
-      setDots(
-        <svg style={{ height: "100%", width: "100%" }}>
-          {Object.keys(actualCoords).map((key, index) => {
-            return (
-              <circle
-                key={`${key}_circle`}
-                id={`${key}_circle`}
-                cx={actualCoords[key].u}
-                cy={actualCoords[key].v}
-                data-location={actualCoords[key]["location"]}
-                r="2"
-                fill="white"
-              />
-            );
-          })}
-        </svg>
-      );
-      // console.log(actualCoords)
-      // setBrainCoords(actualCoords)
-      // setNewBrainCoord(dots.toString())
-    });
-  }, [subject]);
 
-  useEffect(() => {
-    if (dots != undefined) {
-      setNewBrainCoord(JSON.stringify(dots));
-    }
-  }, [dots]);
+      //@ts-ignore
+      setElectrodes(actualCoords)
+    })()
+  }, []);
+
+
 
   return (
-    <>
-      <div
-        id="imgContainer"
-        style={{
-          height: "100%",
-          width: "100%",
-          marginTop: "50px",
-          position: "fixed",
-        }}
-      >
-        {dots}
-      </div>
+    <div
+      style={{ "position": "relative", "display": "inline-block" }}
+      id="container"
+    >
       <Image
-        id="brain"
-        src={brain}
-        style={{ marginTop: "50px", zIndex: -1 }}
-        fluid
-      />
-    </>
+        src={img ? `data:image/jpeg;base64,${window.btoa(img)}` : ''}
+        style={{ "display": "block", "maxWidth": "100%", "height": "auto" }}
+      ></Image>
+      <svg
+        style={
+          {
+            position: "absolute",
+            top: "0",
+            left: "0",
+            width: "100%",
+            height: "100%"
+          }
+        }
+      >
+        {electrodes ?
+          //@ts-ignore
+          Object.keys(electrodes).map(x => {
+            return (<circle
+              //@ts-ignore
+              id={`${x}_circle`}
+              //@ts-ignore
+              cx={electrodes[x].u * (size[0] ? size[0] : document.getElementById('container').offsetWidth)}
+              //@ts-ignore
+              cy={electrodes[x].v * (size[1] ? size[1] : document.getElementById('container').offsetHeight)}
+              fill='green'
+              r='3'
+            ></circle>)
+          })
+
+          : <circle></circle>}
+
+      </svg>
+
+    </div >
   );
 };
